Fixed a few memory issues due to old functions moving data from Qt data to our data structure. Fixed a bug causing spin boxes not to update until window was clicked. Fixed in bug in main testing code that caused color to exceed 255. Fill frame bug still an issue.

+++ GoofyLightsEditor/FrameManipulation.h
@@ -31,6 +31,6 @@ int translateFrame(t_FrameData d, int direction);
 
 // Fills given frame with color
 int fillFrame(t_FrameData &d, QColor rgb_fill);
-int fillFrame2(t_FrameData &d, unsigned short r, unsigned short g, unsigned short b);
+int fillFrame2(t_FrameData &d, int r, int g, int b);
 
 #endif // FRAMEMANIPULATION_H
